Support overriding the getUrl() method in http client transport
This allows the CDN functionality in Indy, built for PyPI support, to work correctly. That functionality requires that the original path be looked up in a database of redirects, then the result used for the actual download URL instead of the one computed from the location URI (the remote INDEX repository, for PyPI) + the path to the file.